Added CSS to the sub pages
I tried to apply some style I've learnt by doing the foundations course
The style of the page isn't that great, but I am very proud. I did it.
It is working. Thats enough for now.
My main learning is, I guess, the importance of classes. And another
learning is definitely, chrome developer tools to inspect the page
are essential.
It makes it easier to understand what CSS is actually doing, when you
see the direct action, when you do change values or properties.